fix eslint
Error: File has too many classes (3). Maximum allowed is 1.  max-classes-per-file
